Move first module to image_cutouts/first
