+++ astroquery/image_cutouts/first/tests/data/image.fits
