+++ astroquery/image_cutouts/first/tests/setup_package.py
@@ -6,4 +6,4 @@ import os
 def get_package_data():
     paths = [os.path.join('data', '*.fits'),
              ]
-    return {'astroquery.first.tests': paths}
+    return {'astroquery.image_cutouts.first.tests': paths}
